requirements.txt
streamlit
biopython
pandas
numpy
plotly
requests
python-dateutil
pytest
fpdf